feat: vote handler (#466)
* feat: created votehandler file to handle votes

* refactor: remove onvote function
